// src/app/modules/project-manager/enum/view.enum.ts
export enum EViewTimeSelect {
  WEEK = 'Week',
  MONTH = 'Month',
  QUATER = 'Quater',
  YEAR = 'Year',
  ALL_TIME = 'All Time',
  CUSTOM_TIME = 'Custom Time',
}

export enum EShortMonthName {
  JAN = 'Jan',
  FEB = 'Feb',
  MAR = 'Mar',
  APR = 'Apr',
  MAY = 'May',
  JUN = 'Jun',
  JUL = 'Jul',
  AUG = 'Aug',
  SEP = 'Sep',
  OCT = 'Oct',
  NOV = 'Nov',
  DEC = 'Dec',
}